cron/src/settings.py
import os
from datetime import datetime

from dotenv import load_dotenv

load_dotenv()

# MySQL credentials
DB_HOST = os.getenv("DB_HOST")
DB_PORT = os.getenv("DB_PORT")
DB_NAME = os.getenv("DB_NAME")
DB_USER = os.getenv("DB_USER")
DB_PASS = os.getenv("DB_PASS")

# Email Server
SMTP_SERVER = os.getenv("SMTP_SERVER")
SMTP_PORT = os.getenv("SMTP_PORT")
SMTP_USER = os.getenv("SMTP_USER")
SMTP_PASS = os.getenv("SMTP_PASS")

# Email settings
EMAIL_FROM = "sender@example.com"
EMAIL_TO = "receiver@example.com"

# S3 settings
S3_BUCKET = os.getenv("S3_BUCKET")
S3_PREFIX = os.getenv("S3_PREFIX")

# S3 credentials
AWS_ACCESS_KEY_ID = os.getenv("AWS_ACCESS_KEY_ID")
AWS_SECRET_ACCESS_KEY = os.getenv("AWS_SECRET_ACCESS_KEY")

NOW = datetime.now().strftime('%Y%m%d_%H%M%S')
BACKUP_FILE = DB_NAME + "_" + NOW + ".sql"
