cron/src/requirements.txt
boto3==1.14.54
botocore==1.17.54
docutils==0.15.2
jmespath==0.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10